Migrate DB into tiny modules
