+++ src/database/_constants.js
@@ -1,0 +1,8 @@
+// Constants that relate to the database modules
+
+module.exports = {
+  defaults: {
+    engine: { atom: "*" },
+    license: "NONE"
+  }
+};
