+++ src/database/getUserByName.js
@@ -1,0 +1,25 @@
+/**
+ * @async
+ * @function getUserByName
+ * @description Get a users details providing their username.
+ * @param {string} username - User name string.
+ * @returns {object} A server status object.
+ */
+
+module.exports = {
+  safe: false,
+  exec: async (sql, username) => {
+    const command = await sql`
+      SELECT * FROM users
+      WHERE username = ${username};
+    `;
+
+    return command.count !== 0
+      ? { ok: true, content: command[0] }
+      : {
+          ok: false,
+          content: `Unable to query for user: ${username}`,
+          short: "not_found",
+        };
+  }
+};
